+++ target_runner_dynamic.py
@@ -12,7 +12,7 @@ with open(instance) as f:
 lbds = [1] * n
 for i in range(n):
     j = int(sys.argv[i*2+6].replace("--lbd", ""))
-    lbd = int(sys.argv[i*2+7])
+    lbd = float(sys.argv[i*2+7])
     lbds[j] = lbd
 
 a, b, c = onell_lambda(n, lbds=lbds, max_evals=bound, seed=seed)
